Fix sending of non-ascii mail
Previously sending mail with non-ascii characters ended up in error
because message was passed as str instance.  However,
smtplib.SMTP.sendmail expects message to be "a string containing
characters in the ASCII range, or a byte string".  Hence, before
calling sendmail check if message can be represented as ASCII and if
not, send byte string encoded with Message.charset (or utf-8).